requirements.txt
fastapi
uvicorn[standard]
python-dotenv
together
gunicorn